Remade Like New
I accidentally updated the wrong version which made me need to merge the new version and older version. I am testing this now but no errors yet! Added more primitives, choices, colors, and a section for GN (Geometry Nodes) additions later. Added a custom object section and a link meshes and Keep object materials or a place to replace them or paint the objects. Added two types of animations Wave and Mouth because the original looked like a mouth in my videos